+++ ty_lib/test_pattern_generator2.py
@@ -28,7 +28,6 @@ import transfer_functions as tf
 import create_gamut_booundary_lut as cgbl
 import font_control as fc
 import color_space as cs
-from common import is_img_shape
 
 CMFS_NAME = 'CIE 1931 2 Degree Standard Observer'
 D65_WHITE = CCS_ILLUMINANTS[CMFS_NAME]['D65']
@@ -2608,6 +2607,160 @@ def make_bt709_hue_chroma_pattern_jzazbz(
     merge(img, img_pat, (0, 0))
 
     return img
+
+
+def line_cross_pattern(nn, num_of_min_line, fg_color, bg_color, mag_rate=1):
+    """
+    Parameters
+    ----------
+    nn : int
+        factor N
+    num_of_min_line : int
+        minimum line number
+    fg_color : ndarray
+        color value. It must be linear.
+    bg_color : ndarray
+        color value. It must be linear.
+    mag_rate : int
+        magnitude rate
+
+    Returns
+    -------
+    ndarray :
+        linear image data.
+    """
+    max_thickness = 2 ** (nn - 1)
+    block_len = max_thickness * num_of_min_line * 2
+    size = block_len * nn
+    print(f"block_len={block_len}, size={size}")
+    img = np.ones((size, size, 3)) * bg_color
+
+    for n_idx in range(nn):
+        thickness = max_thickness // (2 ** n_idx)
+        g_st_pos = [0, block_len * n_idx]
+        num_of_line = num_of_min_line * (2 ** n_idx)
+        for l_idx in range(num_of_line):
+            st_pos = [0, g_st_pos[1] + thickness * 2 * l_idx]
+            # print(f"l_idx={l_idx}, st_pos={st_pos}, thick={thickness}")
+            draw_hv_straight_line(
+                img=img, st_pos=st_pos, width=size, thickness=thickness,
+                color=fg_color, direction='h')
+            draw_hv_straight_line(
+                img=img, st_pos=st_pos, width=size, thickness=thickness,
+                color=fg_color, direction='v')
+
+    img = tf.oetf(img, tf.GAMMA24)
+    out_img = cv2.resize(
+            img, None, fx=mag_rate, fy=mag_rate,
+            interpolation=cv2.INTER_NEAREST)
+    # fname = f"./img/line_cross_nn-{nn}_nol-{num_of_min_line}.png"
+    # write_image(out_img, fname)
+
+    return out_img
+
+
+def draw_hv_straight_line(img, st_pos, width, thickness, color, direction='h'):
+    if direction == 'h':
+        st_pos2 = st_pos
+        ed_pos = [st_pos[0] + width, st_pos[1] + thickness]
+        # print(f"{st_pos2}, {ed_pos}")
+    elif direction == 'v':
+        st_pos2 = [st_pos[1], st_pos[0]]
+        ed_pos = [st_pos2[0] + thickness, st_pos2[1] + width]
+        # print(f"{st_pos2}, {ed_pos}")
+    else:
+        raise ValueError("invalid direction")
+
+    draw_rectangle(img, st_pos2, ed_pos, color)
+
+
+def draw_rectangle(img, st_pos, ed_pos, color):
+    img[st_pos[1]:ed_pos[1], st_pos[0]:ed_pos[0]] = color
+
+
+def draw_border_line(img, st_pos, length, thickness, color):
+    st = st_pos
+    pt1 = [st[0], st[1]]
+    pt2 = [st[0]+length-thickness, st[1]]
+    pt3 = [st[0]+length, st[1]+thickness]
+    pt4 = [st[0], st[1]+length-thickness]
+    pt5 = [st[0]+thickness, st[1]+length]
+    pt6 = [st[0]+length, st[1]+length]
+
+    draw_rectangle(img=img, st_pos=pt1, ed_pos=pt3, color=color)
+    draw_rectangle(img=img, st_pos=pt1, ed_pos=pt5, color=color)
+    draw_rectangle(img=img, st_pos=pt2, ed_pos=pt6, color=color)
+    draw_rectangle(img=img, st_pos=pt4, ed_pos=pt6, color=color)
+
+
+def calc_thickness_for_block(block_idx, num_of_block):
+    bb = block_idx
+    thickness = 2 ** (num_of_block - 1 - bb)
+
+    return thickness
+
+
+def calc_l_for_block(block_idx, num_of_block, num_of_line):
+    bb = block_idx
+    thickness = calc_thickness_for_block(block_idx, num_of_block)
+
+    if bb >= (num_of_block - 1):
+        ll = thickness * num_of_line * 2 * 2 - 1
+    else:
+        ll = thickness * num_of_line * 2 * 2\
+            + calc_l_for_block(
+                block_idx=bb+1, num_of_block=num_of_block,
+                num_of_line=num_of_line)
+
+    return ll
+
+
+def create_multi_border_tp(
+        num_of_line, num_of_block, fg_color, bg_color, mag_rate=1):
+    """
+    num_of_line : int
+        number of the line
+    num_of_block : int
+        number of the block
+    fg_color : ndarray
+        color value. It must be linear.
+    bg_color : ndarray
+        color value. It must be linear.
+    mag_rate : int
+        magnitude rate
+
+    Returns
+    -------
+    ndarray :
+        linear image data
+    """
+    g_st_pos = [0, 0]
+    tp_size = calc_l_for_block(
+        block_idx=0, num_of_block=num_of_block, num_of_line=num_of_line)
+    img = np.ones((tp_size, tp_size, 3)) * bg_color
+    for b_idx in range(num_of_block):
+        thickness = calc_thickness_for_block(b_idx, num_of_block)
+        lb = calc_l_for_block(
+            block_idx=b_idx, num_of_block=num_of_block,
+            num_of_line=num_of_line)
+        for l_idx in range(num_of_line):
+            length = lb - thickness * 2 * 2 * l_idx
+            st_pos_h = g_st_pos[0] + thickness * 2 * l_idx
+            st_pos_v = st_pos_h
+            st_pos = [st_pos_h, st_pos_v]
+            draw_border_line(
+                img=img, st_pos=st_pos, length=length,
+                thickness=thickness, color=fg_color)
+        g_st_pos[0] = g_st_pos[0] + thickness * 2 * num_of_line
+        g_st_pos[1] = g_st_pos[0]
+
+    out_img = cv2.resize(
+            img, None, fx=mag_rate, fy=mag_rate,
+            interpolation=cv2.INTER_NEAREST)
+
+    # write_image(out_img, "./img/multi_border_tp_test.png")
+
+    return out_img
 
 
 if __name__ == '__main__':
